update for testing in ecs

+++ src/main/resources/sql-script/reset-sql.sql
@@ -1,2 +1,3 @@
-drop table product ;
-drop table
+drop table if exists product;
+drop table if exists product_detail;
+drop table if exists "user";
